requirements.txt
joblib==1.2.0
numpy==1.25.2
omegaconf==2.3.0
pandas==1.5.3
PyYAML==6.0.1
scikit_learn==1.2.2
scipy==1.11.4
torch==1.12.1
tqdm==4.65.0
